build: Allow build of shared library
For a shared library you want to set SOVERSION and VERSION for correct
API versioning. This assumes the project uses semantic versioning and
adds the last releases version to the root CMakeLists.txt file. That
version MUST correspond to the Git tag, so it must not be forgotten to
update before a new release!

Signed-off-by: Alexander Dahl <ada@thorsis.com>