platforms: linux: certificates: Add some notes and todos
This (modified) stack builds successfully against OpenSSL 1.0.x, but
fails to so against OpenSSL 1.1.x, because it accesses internals of
OpenSSL which were made opaque with the 1.1.0 release. See their FAQ
and especially the question “My code gets "undefined structure" or
"unknown size" when building with 1.1.0 or later.” [1].

This adds just some notes on research results and TODOs how to proceed.

[1] https://www.openssl.org/docs/faq.html