stackcore: binarydecoder: Use internal logging mechanism
Our commit:c0328167977b broke standalone build of the stack back in
February of 2019.  Later that year we learned how to hook into the
"Trace" mechanism of the stack to let the application log those traces
of the stack.  In commserver that was with v0.0.3-57-g66cc054
(2019-07-08).  So we can use 'OpcUa_Trace()' here to get our information
out, the CommServer is able to display that somehow now, no need to
include header files not part of this stack anymore.

Note: the functional workaround is still in place, this just fixes the
logging part!

Fixes: commit:c0328167977b ("[WORKAROUND] change OpcUa_BinaryDecoder_ReadExtensionObject(...)")